Improved tooltips
git-svn-id: http://www.epigenome.usc.edu/svn/ECControlCenter/branches/gxt@460 d402b20d-0686-0410-8cf3-3d3514b17f74

+++ src/edu/usc/epigenome/eccp/server/ECServiceBackend.java
@@ -439,7 +439,7 @@ public class ECServiceBackend extends RemoteServiceServlet implements ECService
 					 {
 				//		 System.out.println("Numeric metric: "+tname+" "+tvalue);
 						 double dd = Double.valueOf(tvalue);
-						 int n = (int)dd;
+						 long n = (long)dd;
 						 if ((dd-n) == 0 ) {
 							 //System.out.println(itgr.format(n));
 							 p.setValue(itgr.format(n));
